Refactor methods into own files
Refactored the meteor methods into their own files.
Did the same for their tests.
